Review 3: yield-timer leak fix + wait(ms) hardening
Third review round, addressing:

P1 (Codex). HandlerTimer CQE now always decrements pending_ops.
Previously the decrement was gated on yield_armed, but free_conn_impl's
c.reset() clears that flag during close-during-yield deferral; the
target CQE and cancel CQE (which share user_data) also both route here
and each must decrement. Without the fix, slots leaked into pending_free
under repeated close-during-yield.

P2 (Codex). analyze rejects non-wait statements before a wait. The
slice-0 codegen dispatches yields BEFORE the entry block, so the only
valid shape is a (possibly empty) wait prefix followed by the rest of
the route body. `guard ...; wait(50); return 200` would otherwise run
the wait ahead of the guard check. Proper mid-body yields land with
the submit/any/all slice.

P2 (Copilot). handler_gen (u32) per-connection counter replaces
req_start_us for yield-heap stale filtering. req_start_us is μs-truncated
CLOCK_MONOTONIC; under close+reuse churn a new request can legitimately
land in the same microsecond as a stale heap entry. handler_gen bumps
on every on_header_received, persists across Connection::reset, and is
strictly unique per slot lifetime.

P2 (Copilot). Parser widened to u32 end-to-end (AstStatement::status_code,
HirRoute::Wait::ms, MirFunction::Wait::ms). u64 accumulator + UINT32_MAX
cap now allows the full ABI range (~49 days) instead of INT_MAX (~24.8d).

P2 (Copilot). src/jit/codegen.cc now uses
static_cast<u8>(jit::YieldKind::Timer) instead of a hard-coded 3, so an
enum re-order can't silently desync the ABI.

New frontend tests:
- analyze_accepts_wait_near_u32_max    (UINT32_MAX-1 round-trips)
- parser_rejects_wait_above_u32_max    (2^32 rejected)
- analyze_rejects_wait_after_non_wait_statement
- analyze_accepts_wait_as_prefix_with_return

+++ src/compiler/lower_rir.cc
@@ -2886,7 +2886,7 @@ FrontendResult<void> lower_to_rir(const MirModule& mir, FrontendRirModule& out) 
         if (mir.functions[i].waits.len > 0) {
             u32 ms_list[MirFunction::kMaxWaits]{};
             for (u32 wi = 0; wi < mir.functions[i].waits.len; wi++)
-                ms_list[wi] = static_cast<u32>(mir.functions[i].waits[wi].ms);
+                ms_list[wi] = mir.functions[i].waits[wi].ms;
             if (!b.set_yield_payload(fn.value(), ms_list, mir.functions[i].waits.len)) {
                 out.destroy();
                 return frontend_error(FrontendError::OutOfMemory, mir.functions[i].span);
